Prefilter
A new feature "Prefilter" replaces the old "Startup filter" in the Input
settings dialog.

If a prefilter is given, it is applied to each line before processing it
(splitting columns a.s.o), making it fast and memory efficient. (In
contrast, the old "Startup filter" processed the whole file before
applying the filters.)

Using prefilters you can work with files of virtually any size in
Filtus, as long as the prefilter reduces it to a manageable number of
variants.

Example: Suppose you have a whole-genome variant file, with some
annotation indicating the functional type of each variant. A useful
prefilter in this case could be:
"Keep lines which contain 'exonic OR splice'"