[trio] XR inheritance
The previous version of trioRecessiveFilter() did not handle XR
inheritance optimally. Now all homo-/hemizygous variants on X (minus
PAR) are reported if mother is heterozygous and father does not carry
the allele. Note that this is "correct" only if the child is male;
otherwise it requires a de novo event.